Extract implementation
